backend/requirements.txt
flask==3.0.2
flask-cors==4.0.0
sqlalchemy==2.0.25
python-dotenv==1.0.1
netmiko==4.2.0
pysnmp==4.4.12
twilio==9.2.3
requests==2.31.0
